Update header UI
- Add language switcher
- Add none shrink for mode-toggle because it shrinks
- Wrap the seach bar component and update padding
